refactor: moved logger to utils
